refactor: 모임 초대 수락, 거절 로직 gatheringInvitationsWriteService로 이동
