+++ src/domain/services/gathering/gathering-invitations-write.service.ts
@@ -1,7 +1,8 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { ForbiddenException, Inject, Injectable } from '@nestjs/common';
 import { v4 } from 'uuid';
 import { GatheringParticipationEntity } from 'src/domain/entities/gathering/gathering-participation.entity';
 import { GatheringParticipationsRepository } from 'src/domain/interface/gathering/gathering-participations.repository';
+import { FORBIDDEN_MESSAGE } from 'src/domain/error/messages';
 
 @Injectable()
 export class GatheringInvitationsWriteService {
@@ -22,6 +23,30 @@ export class GatheringInvitationsWriteService {
         stdDate,
       ),
     );
+  }
+
+  async accept(invitationId: string, userId: string) {
+    await this.checkIsParticipant(invitationId, userId);
+    await this.gatheringParticipationsRepository.updateStatus(
+      invitationId,
+      'ACCEPTED',
+    );
+  }
+
+  async reject(invitationId: string, userId: string) {
+    await this.checkIsParticipant(invitationId, userId);
+    await this.gatheringParticipationsRepository.delete(invitationId);
+  }
+
+  private async checkIsParticipant(invitationId: string, userId: string) {
+    const participation =
+      await this.gatheringParticipationsRepository.findOneByIdAndParticipantId(
+        invitationId,
+        userId,
+      );
+    if (!participation) {
+      throw new ForbiddenException(FORBIDDEN_MESSAGE);
+    }
   }
 
   async createMany(participations: GatheringParticipationEntity[]) {
